TRAIN-848-849: renaming files for consistency
