+++ src/components/atoms/button/index.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { SearchOutlined } from '@ant-design/icons';
 
-import Button from './Button';
+import { Button } from '.';
 
 export default {
   title: 'Atoms/Button',
